feat: add delete account custom hook
